Refactor the Header, the authReducer, and the Login page
