+++ src/components/Header/routeConstants.ts
@@ -1,0 +1,13 @@
+export default [
+    { name: "Home!", link: "/", activeIndex: 0 },
+    {
+      name: "Sign Up!",
+      link: "/signup",
+      activeIndex: 1
+    },
+    {
+      name: "Login!",
+      link: "/login",
+      activeIndex: 2
+    }
+  ];
