+++ src/pages/Login/styles.ts
@@ -1,0 +1,29 @@
+import { makeStyles, Theme } from "@material-ui/core";
+
+
+export const useStyles = makeStyles((theme: Theme) => ({
+    button: {
+    justifyContent: 'center',
+    height: '45px',
+    "&:hover": {
+      backgroundColor: theme.palette.secondary.dark
+    }
+    },
+    cardContainer: {
+        position: 'absolute',
+        display: "flex",
+        margin: `${theme.spacing(0)} auto`,
+        marginTop: '10em'
+    },
+    input: {
+        color: theme.palette.primary.main
+    },
+    loginCard: {
+        backgroundColor: 'white'
+    },
+    title: {
+        fontSize: 20,
+        textAlign: 'center',
+        color: theme.palette.primary.main
+    }
+}));
