+++ src/redux/actions/authActions.ts
@@ -2,10 +2,10 @@ import axios from 'axios';
 import { Dispatch } from 'react';
 import { ActionCreator } from 'redux';
 import { ThunkAction } from 'redux-thunk';
-import { LoginRequest } from '../reducers/loginRequestReducer';
+import { LoginRequest } from '../reducers/authReducer';
 
-const LOCAL_API_URL = process.env.REACT_APP_API_BASE_URL || 'http://localhost:9191';
-//const CLOUD_API_URL = 'http://saturn-hotdog-super-calculator-backend.cfapps.io';
+// const LOCAL_API_URL = process.env.REACT_APP_API_BASE_URL || 'http://localhost:9191';
+const CLOUD_API_URL = 'http://saturn-hotdog-super-calculator-backend.cfapps.io';
 
 interface ILoginSubmit {
     readonly type: 'LOGIN_SUBMIT';
@@ -17,7 +17,7 @@ interface ILoginComplete {
 }
 
 const apiClient = axios.create({
-    baseURL: LOCAL_API_URL,
+    baseURL: CLOUD_API_URL,
     responseType: 'json',
     headers: {
       'Content-Type': 'application/json'
@@ -29,7 +29,7 @@ const apiClient = axios.create({
   LoginRequest,
   null,
   ILoginSubmit>> = (requestBody: any) => {
-  return (dispatch: Dispatch<LoginRequestActions>) => {
+  return (dispatch: Dispatch<authActions>) => {
         
           return apiClient.post('/auth/login', requestBody)
           .then(({ data }) => {
@@ -46,6 +46,6 @@ const apiClient = axios.create({
   }
 }
 
-export type LoginRequestActions = 
+export type authActions = 
 | ILoginSubmit
 | ILoginComplete
